# docs/swagger.yaml
# SPDX-License-Identifier: AGPL-3.0-or-later
# Copyright (c) 2026 Utkarsh Chourasia

definitions:
  admin.ResetResponse:
    properties:
      contentGroupsDeleted:
        type: integer
      filesDeleted:
        type: integer
      groupMembersDeleted:
        type: integer
      scanSessionsDeleted:
        type: integer
    type: object
  hash.HashProgressResponse:
    properties:
      completedAt:
        type: string
      filesDiscovered:
        type: integer
      filesErrored:
        type: integer
      filesHashed:
        type: integer
      percentComplete:
        type: number
      sessionId:
        type: string
      startedAt:
        type: string
      status:
        type: string
    type: object
  hash.HashStatsResponse:
    properties:
      duplicateFiles:
        type: integer
      groupsWithDupes:
        type: integer
      mastersElected:
        type: integer
      totalFiles:
        type: integer
      totalGroups:
        type: integer
    type: object
  scanner.ScanSession:
    properties:
      completedAt:
        type: string
      errorsEncountered:
        type: integer
      filesDiscovered:
        type: integer
      filesModified:
        type: integer
      filesNew:
        type: integer
      filesSkipped:
        type: integer
      id:
        type: string
      lastError:
        type: string
      rootPaths:
        description: JSONB in DB
        items:
          type: string
        type: array
      startedAt:
        type: string
      status:
        type: string
    type: object
  scans.CleanupOutputResponse:
    properties:
      deletedCount:
        type: integer
      message:
        type: string
    type: object
  scans.FileCountResponse:
    properties:
      totalFiles:
        type: integer
    type: object
  scans.StartScanRequest:
    properties:
      rootPaths:
        items:
          type: string
        type: array
    required:
    - rootPaths
    type: object
  scans.StartScanResponse:
    properties:
      message:
        type: string
      sessionId:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
  description: API documentation for WanderSort
  title: WanderSort API
  version: "1.0"
paths:
  /internal/v1/admin/reset:
    post:
      description: Deletes all scan sessions, file registry entries, content groups
        and group members in a single transaction. Irreversible.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.ResetResponse'
      summary: Reset all application data
      tags:
      - Admin
  /internal/v1/hash/progress:
    get:
      description: Returns files_discovered, files_hashed, files_errored and percent_complete
        for the given session.
      parameters:
      - description: Scan session UUID
        in: query
        name: session_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hash.HashProgressResponse'
      summary: Get hashing progress for a scan session
      tags:
      - Hash
  /internal/v1/hash/stats:
    get:
      description: 'Returns aggregate counts across all content groups: total groups,
        duplicates, masters elected.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hash.HashStatsResponse'
      summary: Get overall content group statistics
      tags:
      - Hash
  /internal/v1/scans/cleanup-output:
    post:
      description: |-
        Checks every ORGANIZED entry in the file registry and removes those whose
        files no longer exist on disk. Does NOT re-index or re-sort any files.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/scans.CleanupOutputResponse'
      summary: Clean up stale organized-library entries
      tags:
      - Scans
  /internal/v1/scans/count:
    get:
      consumes:
      - application/json
      description: Get the total number of files scanned across all sessions
      produces:
      - application/json
      responses:
        "200":
          description: Total file count
          schema:
            $ref: '#/definitions/scans.FileCountResponse'
      summary: Get total file count
      tags:
      - Scans
  /internal/v1/scans/start:
    post:
      consumes:
      - application/json
      description: Start a new scan session with specified root paths
      parameters:
      - description: Start Scan Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/scans.StartScanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/scans.StartScanResponse'
      summary: Start a new scan session
      tags:
      - Scans
  /internal/v1/scans/status:
    get:
      consumes:
      - application/json
      description: Get the status of a scan session
      parameters:
      - in: query
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/scanner.ScanSession'
      summary: Get scan status
      tags:
      - Scans
swagger: "2.0"